# Fix warnings when compiling against jdk11