--- a/akka-actor/src/main/scala/akka/io/Dns.scala
+++ b/akka-actor/src/main/scala/akka/io/Dns.scala
@@ -116,7 +116,11 @@ class DnsExt private[akka] (val system: ExtendedActorSystem, resolverName: Strin
         override def apply(r: String): ActorRef = {
           val settings =
             new Settings(system.settings.config.getConfig("akka.io.dns"), "async-dns")
-          val provider = system.dynamicAccess.getClassFor[DnsProvider](settings.ProviderObjectName).get.newInstance()
+          val provider = system.dynamicAccess
+            .getClassFor[DnsProvider](settings.ProviderObjectName)
+            .get
+            .getDeclaredConstructor()
+            .newInstance()
           system.log.info("Creating async dns resolver {} with manager name {}", settings.Resolver, managerName)
           system.systemActorOf(
             props = Props(
@@ -163,7 +167,11 @@ class DnsExt private[akka] (val system: ExtendedActorSystem, resolverName: Strin
 
   // System DNS resolver
   val provider: DnsProvider =
-    system.dynamicAccess.getClassFor[DnsProvider](Settings.ProviderObjectName).get.newInstance()
+    system.dynamicAccess
+      .getClassFor[DnsProvider](Settings.ProviderObjectName)
+      .get
+      .getDeclaredConstructor()
+      .newInstance()
 
   // System DNS cache
   val cache: Dns = provider.cache