Compiling against jdk11 is possible with #27083