@raboof 2019-06-17
(have to check if performance is on par before merging)