@raboof 2019-07-08
ActorCreationBenchmark looks good but the variance seems rather big (even with 'performance' governor)...
```ActorCreationBenchmark
before:
[info] ActorCreationBenchmark.synchronousStarting    ss    3  200.523 ± 833.761  us/op
[info] ActorCreationBenchmark.synchronousStarting    ss    3  206.496 ± 584.709  us/op
[info] ActorCreationBenchmark.synchronousStarting    ss    3  185.079 ± 325.940  us/op
[info] ActorCreationBenchmark.synchronousStarting    ss    3  226.821 ± 276.221  us/op
after:
[info] ActorCreationBenchmark.synchronousStarting    ss    3  182.120 ± 408.678  us/op
[info] ActorCreationBenchmark.synchronousStarting    ss    3  221.438 ± 216.214  us/op
[info] ActorCreationBenchmark.synchronousStarting    ss    3  162.510 ± 354.119  us/op
[info] ActorCreationBenchmark.synchronousStarting    ss    3  179.318 ± 684.364  us/op
```
Let's be on the safe side and ignore the deprecation for now