@patriknw 2019-07-10
Maybe I was too afraid of performance regression here. Best is to not use the reflection based Props at all, and instead use the lambda/function to create the instance. The only case that doesn't work is for remote deployed actors, and then this small performance concern should matter at all.

Also, Props can be reused and I think we already have some caching in there when reflection is used.