@raboof 2019-07-10
Makes sense. Let's prefer #27308 though.